Hook up GatherContext to the Gather engine
This is a building block for future feature development wherein I
plan to allow things like selective mining or auto-ignore of un-
wanted gathernodes at the player's discretion.  Basically Mobile
stores a GatherContext object and GatherEngine checks this when
the player targets a tile to harvest.  At that point if the
player has a valid context we reuse it, otherwise we fall back to
the old logic of Strike -> BuildNodeList -> IncludeByDistance