Refactor TownCrier API to be simpler
This greatly simplifies the TownCrier logic and reduces the internal
machinery to a simple generic Queue<string> and some supporting
apparatus to handle moving strings into and out of the queue.

Any other module (such as the auction, or minigames) can easily
interface with the TownCrier network by calling
TownCrier.AddAnnoucement( string s ) which will ad the supplied
string to the queue.  On a 1-minute timer, the town criers will
all concurrently say the next item in the queue, Borg=style