Implement the spell Earth Portal
This change adds a Point3D member variable to PlayerMobile which
stores a location when the Earth Portal spell is cast.  Then, on
a second casting, the player is warped to the stored location and
then that location is "erased" by setting it to (0,0,0) upon which
the player must set a new stored location.