Implement Shapeshift
This spell uses a bit of the gump-based logic from Polymorph,
however I've simplified it greatly.  Additionally, the spell
does not check for expansions when determining behaviour.  Instead
it simply takes the Caster's Meditation skill in percentage points
and uses that as the number of seconds, capped at 300 (though the
cap is effectively 150 seconds due to the cap on skills).  Feel
free to change this time limit based on balance and metagame and
whatever makes sense.